requirements.txt
https://www.djangoproject.com/download/1.7b1/tarball/#egg=Django
psycopg2
-e git+https://github.com/opennorth/open511.git#egg=open511
-e git+http://github.com/opennorth/open511-server.git#egg=open511-server
django-extensions==1.2.5

# Uncomment the following line to install the UI
# -e git+http://github.com/opennorth/open511-ui.git#egg=django_open511_ui
